Moved TestEntity and files related to it to the test scope.
