+++ paperlessREST/src/test/java/at/fhtw/swen3/paperless/entities/TestEntity.java
@@ -1,4 +1,4 @@
-package at.fhtw.swen3.paperless.models.entity;
+package at.fhtw.swen3.paperless.entities;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
